ms_game/views.py
"""Define views for MS Game app."""
from rest_framework import mixins, viewsets

from .models import Game
from .serializers import GameSerializer


class GameViewset(
    mixins.CreateModelMixin,
    mixins.RetrieveModelMixin,
    mixins.ListModelMixin,
    viewsets.GenericViewSet,
):
    """
    Viewset for Game model.

    Through this Viewset it is possible to create and retrieve instances.
    """

    queryset = Game.objects.all()
    serializer_class = GameSerializer
