pyproject.toml
[tool.poetry]
name = "minesweeper-api"
version = "0.1.0"
description = ""
authors = ["Federico Jaramillo Martinez <federicojaramillom@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
django_extensions = "^3.0.3"
django-use-email-as-username = "^1.1.1"
psycopg2 = "^2.8.5"
django_heroku = "^0.3.1"
gunicorn = "^20.0.4"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
notebook = "^6.0.3"
bpython = "^0.19"
django-sslserver = "^0.22"
dotenv-cli = "^2.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
